* Modified PhysicalRunway class : Add the calculation code from RunwayWithObstacle. Now RunwayWithObstacle class is no longer working * Modified Test2.java : To test the calculations in the PhysicalRunway class * Modified Runway class : Change the numeric attributes from int to double

+++ src/Model/trunk/SEG/src/model/Runway.java
@@ -3,24 +3,24 @@ package model;
 public class Runway {
 
 	private String name;
-	private int[] TORA, ASDA, TODA, LDA, displacedThreshold;
+	private double[] TORA, ASDA, TODA, LDA, displacedThreshold;
 	
 	//Constructor
-	public Runway(String n, int tora, int asda, int toda, int lda, int disThres){
+	public Runway(String n, double tora, double asda, double toda, double lda, double disThres){
 		name = n;
-		TORA = new int[2];//tora;
+		TORA = new double[2];//tora;
 		TORA[0] = tora;
 		TORA[1] = tora;
-		ASDA = new int[2];//asda;
+		ASDA = new double[2];//asda;
 		ASDA[0] = asda;
 		ASDA[1] = asda;
-		TODA = new int[2];//toda;
+		TODA = new double[2];//toda;
 		TODA[0] = toda;
 		TODA[1] = toda;
-		LDA = new int[2];//lda;
+		LDA = new double[2];//lda;
 		LDA[0] = lda;
 		LDA[1] = lda;
-		displacedThreshold = new int[2];//disThres;
+		displacedThreshold = new double[2];//disThres;
 		displacedThreshold[0] = disThres;
 		displacedThreshold[1] = disThres;
 	}
@@ -34,43 +34,43 @@ public class Runway {
 		this.name = name;
 	}
 
-	public int getTORA(int option) {
+	public double getTORA(int option) {
 		return TORA[option];
 	}
 
-	public void setTORA(int option, int tORA) {
+	public void setTORA(int option, double tORA) {
 		TORA[option] = tORA;
 	}
 
-	public int getASDA(int option) {
+	public double getASDA(int option) {
 		return ASDA[option];
 	}
 
-	public void setASDA(int option, int aSDA) {
+	public void setASDA(int option, double aSDA) {
 		ASDA[option] = aSDA;
 	}
 
-	public int getTODA(int option) {
+	public double getTODA(int option) {
 		return TODA[option];
 	}
 
-	public void setTODA(int option, int tODA) {
+	public void setTODA(int option, double tODA) {
 		TODA[option] = tODA;
 	}
 
-	public int getLDA(int option) {
+	public double getLDA(int option) {
 		return LDA[option];
 	}
 
-	public void setLDA(int option, int lDA) {
+	public void setLDA(int option, double lDA) {
 		LDA[option] = lDA;
 	}
 
-	public int getDisplacedThreshold(int option) {
+	public double getDisplacedThreshold(int option) {
 		return displacedThreshold[option];
 	}
 
-	public void setDisplacedThreshold(int option, int displacedThreshold) {
+	public void setDisplacedThreshold(int option, double displacedThreshold) {
 		this.displacedThreshold[option] = displacedThreshold;
 	}
